refactor: rename latest-news.ts to informasi-terbaru.ts
