+++ __tests__/pages/index.test.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-import { latestNews } from "~/lib/content/__mocks__/builders/latest-news";
+import { latestNews } from "~/lib/content/__mocks__/builders/informasi-terbaru";
 import HomePage from "~/pages";
 
 import { render, screen } from "@testing-library/react";
