--- lib/content/latest-news.ts
@@ -1,10 +1,0 @@
-type LatestNewsItemAttributes = {
-  title: string;
-  date: string;
-  description?: string;
-} & ({ link: string; link_text: string } | { link?: never });
-
-export interface LatestNewsItem {
-  attributes: LatestNewsItemAttributes;
-  html: string;
-}
